requirements.txt
torch==2.0.1
torchvision==0.15.2
onnx==1.14.0
onnxruntime-gpu==1.15.1
Pillow==9.5.0
numpy==1.26.4
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1